requirements.txt
aws-cdk.core
aws-cdk.aws_lambda
aws-cdk.aws_iam
aws-cdk.aws_secretsmanager
botocore